+++ components/calendar/src/types.rs
@@ -759,73 +759,3 @@ impl From<usize> for IsoWeekday {
         unsafe { core::mem::transmute(ordinal) }
     }
 }
-
-/// A moment is a RataDie with a fractional part giving the time of day.
-///
-/// NOTE: This should not cause overflow errors for most cases, but consider
-/// alternative implementations if necessary.
-#[derive(Debug, Copy, Clone, PartialEq, PartialOrd)]
-#[doc(hidden)] // This type is unstable
-pub struct Moment(f64);
-
-/// Add a number of days to a Moment
-impl Add<f64> for Moment {
-    type Output = Self;
-    fn add(self, rhs: f64) -> Self::Output {
-        Self(self.0 + rhs)
-    }
-}
-
-impl AddAssign<f64> for Moment {
-    fn add_assign(&mut self, rhs: f64) {
-        self.0 += rhs;
-    }
-}
-
-/// Subtract a number of days from a Moment
-impl Sub<f64> for Moment {
-    type Output = Self;
-    fn sub(self, rhs: f64) -> Self::Output {
-        Self(self.0 - rhs)
-    }
-}
-
-impl SubAssign<f64> for Moment {
-    fn sub_assign(&mut self, rhs: f64) {
-        self.0 -= rhs;
-    }
-}
-
-/// Calculate the number of days between two moments
-impl Sub for Moment {
-    type Output = f64;
-    fn sub(self, rhs: Self) -> Self::Output {
-        self.0 - rhs.0
-    }
-}
-
-impl Moment {
-    /// Create a new moment
-    pub const fn new(value: f64) -> Moment {
-        Moment(value)
-    }
-
-    /// Get the inner field of a Moment
-    pub const fn inner(&self) -> f64 {
-        self.0
-    }
-
-    /// Get the RataDie of a Moment
-    pub(crate) fn as_rata_die(&self) -> RataDie {
-        RataDie::new(libm::floor(self.0) as i64)
-    }
-}
-
-#[test]
-fn test_moment_to_rata_die_conversion() {
-    for i in -1000..=1000 {
-        let moment = Moment::new(i as f64);
-        let rata_die = moment.as_rata_die();
-        assert_eq!(rata_die.to_i64_date(), i);
-    }
-}
